Moved CommonStyles components.
