+++ frontend/src/components/welcome/how-it-works-people.js
@@ -13,7 +13,7 @@ import WorkIcon from 'material-ui-icons/Work'
 import AccountBalanceWalletIcon from 'material-ui-icons/AccountBalanceWallet'
 import Transition from '../transition'
 
-import { InfoList, MainTitle } from './CommonStyles'
+import { InfoList, MainTitle } from './components/CommonStyles'
 
 class HowItWorksPeople extends Component {
   constructor (props) {
